chore(labkit): 把共用的 WASM 资源挪到中立目录，为 gpulab 的 CUDA 语法腾位置
上一片把 shell 的解析器移进了 labkit，但它仍然从 `/opslab/tree-sitter-bash.wasm`
取语法 —— 一个两个实验台共用的模块，去 opslab 的目录里拿东西。gpulab 接下来要
加 tree-sitter-cuda.wasm，再往 public/opslab/ 里塞就彻底拧了。

按归属分目录：

  public/labkit/web-tree-sitter.wasm    两个实验台共用的 tree-sitter 运行时
  public/labkit/tree-sitter-bash.wasm   shell 语法，labkit
  public/gpulab/                        gpulab 专用（下一片开始用）
  public/opslab/opslab-cli.wasm         Go 编出来的多合一 CLI，不动

`scripts/copy-opslab-assets.js` 一并改名成 `copy-lab-assets.js` —— 它现在管的
不只是 opslab。

electron-builder 那边的 `files` 与 `asarUnpack` 都是**逐个文件**的白名单
（那个白名单是有来历的：通配符曾经把 120MB 的旧 kubectl.wasm 打进过包），
所以两处都手工加了新路径，并在脚本与配置里互相留了「改一处要改另一处」的注释。

release.yml 不用动：这两个文件来自 node_modules，由 prebuild 钩子拷过去，
而 release.yml 本来就跑 npm run build。只有 Go 编的 opslab-cli.wasm 需要
release.yml 里那条独立的 build-wasm job。

验证不止于「本地能跑」：

  - npx electron-builder --dir 出真实产物，asar 里两个文件都在，
    asarUnpack 也解出来了，且 public/ 下没有任何旧的 opslab wasm 残留
  - next start 跑生产构建，/labkit/*.wasm 两个都是 HTTP 200 且字节数与
    磁盘文件完全一致；旧的 /opslab/tree-sitter-bash.wasm 已经 404
  - tests/labkit/machine.test.ts 里那条断言同步改到新路径 —— 它验证的正是
    「代码请求的 URL」与「拷贝脚本的落点」一致，是这条管线的守卫

npm test 9/9、projects:verify 全绿、tsc 干净。
